Inserted bond weight for B-C
I added the bond weight matrix for the B-C contraction,
as well as updating the bond weights list for the two
updated directions.  I have to figure out when to include
the bond weights in the squeezer update and when not to,
as well as to make sure I use the un-updated bond weights
before updating the tensors/bond weight matrices.